move isAuth middleware to src folder and update imports
